feat: Add Experiences page
The code changes add a new page called "Experiences" to the project. This page displays the user's resume using the `Resume` component. The page is wrapped in a `PublicLayout` component and includes a `Head` component to set the page title.

Note: The commit message has been generated based on the provided code changes and recent commits.